% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_province_codes}
\alias{convert_province_codes}
\title{Convert Canadian Province Codes to Full Names}
\usage{
convert_province_codes(codes)
}
\arguments{
\item{codes}{Character vector of province/territory codes}
}
\value{
Character vector of full province/territory names. Unknown codes
return NA.
}
\description{
Converts Canadian province/territory codes to their full names using
a standardized lookup table.
}
\examples{
convert_province_codes(c("ON", "QC", "BC"))
convert_province_codes(c("ON", "XX", "AB"))  # Unknown code returns NA
}
