% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barque_output_loading.R
\name{load_barque_data}
\alias{load_barque_data}
\title{Load and Process Barque Output Data}
\usage{
load_barque_data(barque_output_folder, samples_ids)
}
\arguments{
\item{barque_output_folder}{Path to the barque output folder}

\item{samples_ids}{Vector of sample IDs to include in the analysis}
}
\value{
A list containing single_hits and multi_hits data frames
}
\description{
This function loads the species table from barque output and processes it
by removing totals and separating single hits from multi-hits.
}
\examples{
\dontrun{
data <- load_barque_data("/path/to/barque/output", c("ST1", "ST2", "ST3"))
}
}
