% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{province_codes}
\alias{province_codes}
\title{Canadian Province Code Lookup Table}
\format{
A named character vector with 14 elements:
\describe{
\item{ON}{Ontario}
\item{QC}{Quebec}
\item{LB}{Labrador}
\item{NB}{New Brunswick}
\item{NT}{Northwest Territories}
\item{NF}{Newfoundland}
\item{YT}{Yukon}
\item{PE}{Prince Edward Island}
\item{NU}{Nunavut}
\item{SK}{Saskatchewan}
\item{MB}{Manitoba}
\item{NS}{Nova Scotia}
\item{AB}{Alberta}
\item{BC}{British Columbia}
}
}
\usage{
province_codes
}
\description{
A named vector mapping Canadian province/territory codes to full names.
}
\keyword{datasets}
